backend/requirements.txt
Django==5.2.1
celery==5.5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-yasg==1.21.10

# Celery + Django integration
django-celery-beat==2.8.1
django-celery-results==2.6.0

# PostgreSQL driver
psycopg2-binary==2.9.10

python-decouple==3.8
